requirements.txt
streamlit
pandas
snowflake-snowpark-python
plotly
altair
openpyxl